refactor(mentor): 지난 시간 계산 함수 util 파일에 리팩토링
